Aktualizacja o nową klasę CalculatorAdvanced, dodanie do metod parametrów
