+++ Program.cs
@@ -25,12 +25,15 @@ namespace FirstProject
                     + "*  Minimum value from the table \n");
 
                 string operation = Console.ReadLine();
-                Calculator calculator = new Calculator();
 
                 Console.WriteLine("Podaj dwie liczby: ");
                 int argument1 = Convert.ToInt32(Console.ReadLine());
                 int argument2 = Convert.ToInt32(Console.ReadLine());
 
+
+                
+                Calculator calculator = new Calculator();
+                CalculatorAdvanced calculatoradvanced = new CalculatorAdvanced();
 
                 if (operation == "1")
                 {
@@ -38,43 +41,43 @@ namespace FirstProject
                 }
                 else if (operation == "2")
                 {
-                    Console.WriteLine(calculator.Subtraction());
+                    Console.WriteLine(calculator.Subtraction(argument1, argument2));
                 }
                 else if (operation == "3")
                 {
-                    Console.WriteLine(calculator.Multiplication());
+                    Console.WriteLine(calculator.Multiplication(argument1, argument2));
                 }
                 else if (operation == "4")
                 {
-                    Console.WriteLine(calculator.Division());
+                    Console.WriteLine(calculator.Division(argument1, argument2));
                 }
                 else if (operation == "5")
                 {
-                    Console.WriteLine(calculator.PrimeNumbers());
+                    Console.WriteLine(calculatoradvanced.PrimeNumbers());
                 }
                 else if (operation == "6")
                 {
-                    Console.WriteLine(calculator.Factorial());
+                    Console.WriteLine(calculatoradvanced.Factorial());
                 }
                 else if (operation == "7")
                 {
-                    Console.WriteLine(calculator.PerfectNumber());
+                    Console.WriteLine(calculatoradvanced.PerfectNumber());
                 }
                 else if (operation == "8")
                 {
-                    Console.WriteLine(calculator.ToThePower());
+                    Console.WriteLine(calculator.ToThePower(argument1, argument2));
                 }
                 else if (operation == "9")
                 {
-                    Console.WriteLine(calculator.AverageNumberFromTheTable());
+                    Console.WriteLine(calculatoradvanced.AverageNumberFromTheTable());
                 }
                 else if (operation == "0")
                 {
-                    Console.WriteLine(calculator.MaximumValueFromTheTable());
+                    Console.WriteLine(calculatoradvanced.MaximumValueFromTheTable());
                 }
                 else if (operation == "*")
                 {
-                    Console.WriteLine(calculator.MinimumValueFromTheTable());
+                    Console.WriteLine(calculatoradvanced.MinimumValueFromTheTable());
                 }
 
                 Console.ReadLine();
